+++ JPAHibernate/src/main/java/com/sparta/jpahibernate/controllers/EmployeeController.java
@@ -5,6 +5,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.json.JsonMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import com.sparta.jpahibernate.dao.concretes.EmployeeDAOImpl;
+import com.sparta.jpahibernate.dao.concretes.UserDAOImpl;
 import com.sparta.jpahibernate.dto.EmployeeDTO;
 import com.sparta.jpahibernate.dto.EmpsByDeptsDTO;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -25,32 +26,46 @@ public class EmployeeController {
     @Autowired
     private EmployeeDAOImpl empDao;
 
+    @Autowired
+    private UserDAOImpl userDao;
+
     @PostMapping("/add")
-    public ResponseEntity<EmployeeDTO> addNewEmployee(@RequestBody EmployeeDTO newEmployee){
-        empDao.save(newEmployee);
-        return new ResponseEntity<>(newEmployee, HttpStatus.OK);
+    public ResponseEntity<String> addNewEmployee(@RequestBody EmployeeDTO newEmployee, @RequestParam String apiKey){
+        if (userDao.isAdmin(apiKey) || userDao.isUpdate(apiKey)){
+            empDao.save(newEmployee);
+            return new ResponseEntity<>(newEmployee.toString(), HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>("\"message\":\"User could not be added\"}",
+                    HttpStatus.UNAUTHORIZED);
+        }
+
     }
 
     @PutMapping("/update/create")
-    public ResponseEntity<EmployeeDTO> updateEmployeePut(@RequestBody EmployeeDTO newState){
-        Optional<EmployeeDTO> original = empDao.findById(newState.getId());
-        if (original.isPresent()) {
-            String[] params = new String[]{newState.getId().toString(), newState.getBirthDate().toString(), newState.getFirstName(), newState.getLastName(), newState.getGender(), newState.getHireDate().toString()};
-            empDao.update(newState.getId(), params);
-            return new ResponseEntity<>(empDao.findById(newState.getId()).get(), HttpStatus.OK);
+    public ResponseEntity<EmployeeDTO> updateEmployeePut(@RequestBody EmployeeDTO newState, @RequestParam String apiKey){
+        if (userDao.isAdmin(apiKey) || userDao.isUpdate(apiKey)) {
+            Optional<EmployeeDTO> original = empDao.findById(newState.getId());
+            if (original.isPresent()) {
+                String[] params = new String[]{newState.getId().toString(), newState.getBirthDate().toString(), newState.getFirstName(), newState.getLastName(), newState.getGender(), newState.getHireDate().toString()};
+                empDao.update(newState.getId(), params);
+                return new ResponseEntity<>(empDao.findById(newState.getId()).get(), HttpStatus.OK);
+            } else {
+                empDao.save(newState);
+                return new ResponseEntity<>(empDao.findById(newState.getId()).get(), HttpStatus.OK);
+            }
         }
-        else {
-            empDao.save(newState);
-            return new ResponseEntity<>(empDao.findById(newState.getId()).get(), HttpStatus.OK);
-        }
+        return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
     }
 
     @PatchMapping("/update")
-    public ResponseEntity<EmployeeDTO> updateEmployeePatch(@RequestBody EmployeeDTO newState){
-        String[] params = new String[] {newState.getId().toString(), newState.getBirthDate().toString(), newState.getFirstName(), newState.getLastName(), newState.getGender(), newState.getHireDate().toString()};
-        empDao.update(newState.getId(), params);
-        System.out.println(empDao.findById(newState.getId()).get());
-        return new ResponseEntity<>(empDao.findById(newState.getId()).get(), HttpStatus.OK);
+    public ResponseEntity<EmployeeDTO> updateEmployeePatch(@RequestBody EmployeeDTO newState, @RequestParam String apiKey){
+        if (userDao.isAdmin(apiKey) || userDao.isUpdate(apiKey)) {
+            String[] params = new String[]{newState.getId().toString(), newState.getBirthDate().toString(), newState.getFirstName(), newState.getLastName(), newState.getGender(), newState.getHireDate().toString()};
+            empDao.update(newState.getId(), params);
+            System.out.println(empDao.findById(newState.getId()).get());
+            return new ResponseEntity<>(empDao.findById(newState.getId()).get(), HttpStatus.OK);
+        }
+        return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
     }
 
     @GetMapping("/{id}")
@@ -81,10 +96,12 @@ public class EmployeeController {
 //    }
 
     @DeleteMapping("/delete/{id}")
-    public void deleteById(@PathVariable int id){
-        empDao.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND,
-                "No employee with the specific ID could be found"));
-        empDao.deleteById(id);
+    public void deleteById(@PathVariable int id, @RequestParam String apiKey){
+        if (userDao.isAdmin(apiKey)) {
+            empDao.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND,
+                    "No employee with the specific ID could be found"));
+            empDao.deleteById(id);
+        }
     }
 
 //    @DeleteMapping("/{id}")
@@ -102,7 +119,8 @@ public class EmployeeController {
 
     @GetMapping("/countNumberOfEmployeesLeftDepartmentByYear")
     public int countNumberOfEmployeesLeftDepartmentByYear(@RequestParam("dep") String dep,
-                                                          @RequestParam("year") String year){
+                                                          @RequestParam("year") String year,
+                                                          @RequestParam String apiKey){
 
         System.out.println( "In the year:" + year + "the number of employees left in department:" + dep );
         int res = empDao.countNumberOfEmployeesLeftDepartmentByYear(dep,year);
@@ -113,13 +131,15 @@ public class EmployeeController {
     @GetMapping("/findByDepartmentAndDate")
     public List<EmpsByDeptsDTO> findByDepartmentAndDate(@RequestParam("dep") String dep,
                                                         @RequestParam("fromDate") LocalDate fromDate,
-                                                        @RequestParam("toDate") LocalDate toDate){
+                                                        @RequestParam("toDate") LocalDate toDate,
+                                                        @RequestParam String apiKey){
 
         return empDao.findByDepartmentAndDate(dep, fromDate, toDate);
     }
 
     @GetMapping("/findByLastName")
-    public List<EmployeeDTO> findByLastName(@RequestParam("lastName") String lastName){
+    public List<EmployeeDTO> findByLastName(@RequestParam("lastName") String lastName,
+                                            @RequestParam String apiKey){
         return empDao.findByLastName(lastName);
     }
 }
